pyproject.toml
[project]
name = "stackoverflow-mcp"
version = "0.1.0"
description = "StackOverflow MCP server for querying programming Q&A with enhanced CLI and auto-configuration"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
maintainers = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = [
    "stackoverflow", 
    "mcp", 
    "model-context-protocol", 
    "api", 
    "programming", 
    "qa",
    "search",
    "cli",
    "server"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Documentation",
    "Topic :: Text Processing :: Markup :: Markdown"
]
dependencies = [
    "mcp>=1.0.0",
    "fastmcp>=0.9.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.11.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/stackoverflow-mcp"
Documentation = "https://github.com/yourusername/stackoverflow-mcp#readme"
Repository = "https://github.com/yourusername/stackoverflow-mcp.git"
"Bug Tracker" = "https://github.com/yourusername/stackoverflow-mcp/issues"
Changelog = "https://github.com/yourusername/stackoverflow-mcp/blob/main/CHANGELOG.md"

[project.scripts]
stackoverflow-mcp = "stackoverflow_mcp.main:main"
stackoverflow-mcp-fastmcp = "stackoverflow_mcp.fastmcp_main:main"
stackoverflow-mcp-server = "stackoverflow_mcp.main:main"
stackoverflow-mcp-fastmcp-server = "stackoverflow_mcp.fastmcp_main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-httpx>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "coverage>=7.0.0"
]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-httpx>=0.21.0",
    "coverage>=7.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stackoverflow_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
    "/CHANGELOG.md"
]

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"tests/**/*" = ["B011"]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
pythonpath = [
    "src"
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/stackoverflow_mcp"]
omit = [
    "tests/*",
    "src/stackoverflow_mcp/__pycache__/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:"
]
